Add files via upload
The statistics include degree and distribution of nodes, cliques, diameter, and measures of centrality. Appropriate network visualization techniques to will be used to highlight community structures. For community detection, several algorithms such as, GN algorithm, modularity method, leading eigenvector method and label propagation method will be applied. In addition to community detection, random graphs, such as the Stochastic block model will be generated so as to analyze the existence and detectability of community structure in networks with varying degrees of randomness. Finally, spectral test will be applied to determine the hypothesis of accepting or rejecting the existence of community in the network structure.